// src/types.ts
export type Priority = 'low' | 'medium' | 'high';

export interface Task {
  id: string;
  title: string;
  time: string;
  duration?: string;
  completed: boolean;
  completedAt?: string;
  priority: Priority;
  dueDate?: string;
  createdAt: string;
  focusTimeSpent: number; // in seconds
}